pyproject.toml
[project]
name = "streamlit-crewai-test1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.118.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
]
